Created PokeAPI to pokemon helper
- Added PokeAPI to pokemon handler with associated interface
- Moved IPokemonHandler into handler directory
