+++ PokedexAPI/PokedexAPI/Interfaces/Helpers/IPokeApiToPokemonHelper.cs
@@ -1,0 +1,9 @@
+﻿using PokedexAPI.Models;
+
+namespace PokedexAPI.Interfaces.Helpers
+{
+    public interface IPokeApiToPokemonHelper
+    {
+        public Pokemon ConvertPokeApiResponseToPokemon(string pokemon, string repsonse);
+    }
+}
